Pipeline support for tokens and accessors
Adds pipeline support for all functions that consume or return tokens or accessors.
Added logic to handle a soft-failure scenario that occurs when an attempt is made to update the lease/ttl on an LDAP-based token.
Adds two new private functions to parse a token/accessor object, which can take the form of a JSON snippet, PSObject or String, and return just the string containing the token or accessor.