Vault Status BugFixes and New Function
Fixes a misleading error that will be thrown when all vault nodes are sealed, but can also be thrown when Consul DNS has not yet failed over to the other Vault Node.
New function Test-VaultHealth which returns a response code and status information about a queried node.
Fixes a bug where Test-VaultSessionVariable would be called even in scenarios when VaultNodeOverride parameterset is being used. This defeats the purpose of VaultNodeOverride, as its generally understood that vault is sealed when attempting to Unprotect or potentially sealed when attempting to query health information.