Hashtable OutputType option
Adds support for all function that return output to return the output as a Hashtable.
Adds ConvertTo-Hashtable to convert data to hashtables. Nested information inside PSObjects/Json are also converted to hashtables.
Adds -JustData parameter to Get-VaultLoginToken.
Adds Hashtable choice to all ValidateSets for OutputType.
Addresses a bug where output would not be returned to the console if JustData were specified in Get-VaultWrapping and the specified token was a Vault-wrapped token.