ConvertTo-PSCredential function & Bugfixes
New function ConvertTo-PSCredential consumes the result of Get-VaultKVSecret or Get-VaultCubbyholeSecret and returns a PSCredential.
Fixes a bug where no output would be returned if -JustData were specified in Get-VaultCubbyholeSecret.
New function ConvertFrom-Hashtable consumes a hashtable and converts it to a PSObject.
Adds ConvertFrom-Hashtable and ConvertTo-PSCredential to list of exported functions in psVaultUtils.psd1.